refactor: 상품 탭 관련 타입 파일 분리
